@@ -1,26 +1,31 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.17'
+gem 'rails', '4.0.13'
+gem 'protected_attributes'
 gem 'color-tools', '>= 1.3.0'
 gem "will_paginate", "~> 3.0"
 gem "mysql2", "~> 0.3.11"
 gem 'dynamic_form'
 gem 'railroady'
-gem "therubyracer", "~> 0.11.4"
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
 gem 'flot-rails'
-gem 'thinreports', git: 'https://github.com/thinreports/thinreports-generator'
+gem 'thinreports'
 gem 'thinreports-rails'
 gem 'ruby-units'
 gem 'gnuplot'
 
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails'
-  gem 'coffee-rails'
-  gem 'uglifier'
+gem 'pry-rails'
+gem 'sass-rails'
+gem 'coffee-rails'
+gem 'uglifier'
+
+group :development do
+  #gem 'better_errors'
+  #gem 'binding_of_caller'
+  gem 'quiet_assets'
+  #gem 'rails_layout'
+  #gem 'autoprefixer-rails'
 end
 
 gem 'passenger'
