@@ -3,13 +3,13 @@
 class IngredientsRecipesController < ApplicationController
   def index
     @recipe = Recipe.find(params[:recipe_id], :include=>'ingredient_recipe')#, :order=>'ingredients_recipes.id desc')
-    @ingredients = Ingredient.find :all
+    @ingredients = Ingredient.all
   end
 
   def create
     unless params[:ingredient_recipe][:code].blank?
       @recipe = Recipe.find(params[:recipe_id], :include=>'ingredient_recipe')
-      ingredient = Ingredient.find_by_code(params[:ingredient_recipe][:code])
+      ingredient = Ingredient.where(code: params[:ingredient_recipe][:code]).first
 
       ingredient_recipe = IngredientRecipe.new
       ingredient_recipe.ingredient = ingredient
@@ -37,7 +37,7 @@ class IngredientsRecipesController < ApplicationController
   def destroy
     @ingredient_recipe = IngredientRecipe.find params[:id]
     @ingredient_recipe.eliminate
-    if @ingredient_recipe.errors.size.zero?
+    if @ingredient_recipe.errors.empty?
       flash[:notice] = "Ingrediente eliminado de la receta con éxito"
     else
       logger.error("Error eliminando ingrediente de la receta: #{@ingredient_recipe.errors.inspect}")
