@@ -91,7 +91,7 @@ class ProductLotsController < ApplicationController
     if product_lot_transactions.none?
       @lot = ProductLot.find params[:id]
       @lot.eliminate
-      if @lot.errors.size.zero?
+      if @lot.errors.empty?
         flash[:notice] = "Lote eliminado con éxito"
       else
         logger.error("Error eliminando lote: #{@lot.errors.inspect}")
