@@ -33,7 +33,7 @@ class SchedulesController < ApplicationController
   def destroy
     @schedule = Schedule.find params[:id]
     @schedule.eliminate
-    if @schedule.errors.size.zero?
+    if @schedule.errors.empty?
       flash[:notice] = "Turno eliminado con éxito"
     else
       logger.error("Error eliminando turno: #{@schedule.errors.inspect}")
