@@ -21,8 +21,8 @@ class PermissionRole < ActiveRecord::Base
   def self.generate_scada_role
     scada_role = Role.find(2)
     scada_role.permission_role.clear
-    permissions = Permission.find :all, :conditions=>["module = 'batches'"]
-    permissions << Permission.find(:first, :conditions=>["module = 'transactions' AND action = 'SA-CSM'"])
+    permissions = Permission.where(["module = 'batches'"])
+    permissions << Permission.where(["module = 'transactions' AND action = 'SA-CSM'"]).first
     permissions.each do |perm|
       perm_role = PermissionRole.new
       perm_role.permission_id = perm.id
